% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrasp.create.R
\docType{class}
\name{ggrasp-class}
\alias{ggrasp-class}
\title{An S4 class representing the GGRaSP data and output}
\description{
An S4 class representing the GGRaSP data and output
}
\section{Slots}{

\describe{
\item{\code{dist.mst}}{The distance matrix showing the distances between different genomes}

\item{\code{phy}}{The phylogenetic tree in newick format}

\item{\code{rank}}{The ranks of the respective genomes with lower getting higher priority in being called as a medoid}

\item{\code{cluster}}{A vector giving the numeric cluster ID for each genome}

\item{\code{h}}{The threshold variable used to make the clusters}

\item{\code{medoids}}{A vector giving the medoid for each cluster}

\item{\code{gmm}}{A data.frame containing all the gaussian distributions used to find the threshold when available}

\item{\code{gmm.orig}}{A data.frame containing all the gaussian distributions prior to cleaning. Used to recalculate the threshold when needed}
}}

