% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrasp.cluster.R
\name{ggrasp.addRanks}
\alias{ggrasp.addRanks}
\title{ggrasp.addRanks}
\usage{
ggrasp.addRanks(x, rank.file)
}
\arguments{
\item{x}{the GGRaSP object for which the ranks will be added.}

\item{rank.file}{string pointing to a file containing the ranks}
}
\value{
A GGRaSP object where the ranks have been entirely redefined with the ranks in rank.file
}
\description{
adds a rank file to a GGRaSP object. If clusters have been defined, the medoids will be re-defined
}
