% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrasp.create.R
\name{ggrasp.load}
\alias{ggrasp.load}
\title{ggrasp.load}
\usage{
ggrasp.load(file, file.format, rank.file, offset, tree.method = "complete")
}
\arguments{
\item{file}{File containing the tree, matrix or sequence alignment used to initialize the ggrasp object. Required.}

\item{file.format}{The format the file is in, with tree, fasta and matrix accepted. If not given the program makes a guess.}

\item{rank.file}{File containing the ranks of genomes in a tab-delineated file with the genome in column 1 and the rank in column 2. The rank is a non-negative number.}

\item{offset}{Numeric representing a perfect match. Default is 0.}

\item{tree.method}{The method used to make the tree from a distance matrix. "Complete" (Default), "Average", "Single", and "nj" (Neighbor Joining) are currently available.}
}
\value{
Returns a class GGRaSP variable
}
\description{
ggrasp.load() initializes a class GGRaSP object from a file containing either a tree, a distance matrix or a multi-fasta alignment. The returned object can subsequently be clustered using ggrasp.cluster().
}
\examples{
#The following data is from Chavda et al 2016 which phylotyped Enterobacter genomes
# Our example uses the data underpinning the tree shown in Figure 2
# Also included is a ranking file to prioritize closed Enterobactor genomes

library(ggrasp);
Enter.tree <- ggrasp.load(system.file("extdata", "Enter.kSNP.tree", package="ggrasp"), file.format = "tree", rank.file =system.file("extdata", "Enter.kSNP.ranks", package="ggrasp"));

# Other options include loading by fasta file:
Enter.tree <- ggrasp.load(system.file("extdata", "Enter.kSNP2.fasta", package="ggrasp"), file.format = "fasta", rank.file =system.file("extdata", "Enter.kSNP.ranks", package="ggrasp"))

# and by distance matrix. Since this distance matrix is actually percent identity, we will us an offset of 100
Enter.tree <- ggrasp.load(system.file("extdata", "Enter.ANI.mat", package="ggrasp"), file.format = "matrix", rank.file =system.file("extdata", "Enter.kSNP.ranks", package="ggrasp"), offset = 100)

# Use summary() to examine the data loaded
summary(Enter.tree)

#Use plot() to see the tree
plot(Enter.tree)
}
