% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.R
\name{nest1_.Spatial}
\alias{nest1_.Spatial}
\alias{nest_.Spatial}
\title{Nested Spatial}
\usage{
nest1_.Spatial(data, ...)

nest_.Spatial(data, ...)
}
\arguments{
\item{data}{sp Spatial object}

\item{...}{ignored}
}
\value{
nested tibble
}
\description{
Create nested tables from Spatial classes.
}
\details{
For \code{nest_} the tables are doubly nested.
For \code{nest1_} the tables are singly nested.
}
\examples{
library(tidyr)
data(holy_poly)
library(dplyr)
spdata <- spbabel::sp(holy_poly, attr_tab = data_frame(mydat = c("a", "b")))
##ggplot(holy_poly, aes(x = x_, y = y_)) + geom_holygon(aes(group = branch_, fill = object_))
ggg <- nest(spdata)
plot(ggg, col = rainbow(nrow(ggg), alpha = 0.5))
## plot with ggvis
##holy_poly  \%>\% group_by(object_)  \%>\% ggvis(~x_, ~y_, fill = ~object_, stroke = ~branch_)  \%>\% layer_paths()

}

