% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.ggrasp}
\alias{plot.ggrasp}
\title{plot.ggrasp}
\usage{
\method{plot}{ggrasp}(x, type, layout = "circular", ...)
}
\arguments{
\item{x}{ggrasp-class object to be plotted}

\item{type}{Type of plot generated, either "tree" (the full tree with the clusters shown as grouped leaves), "reduced" (tree with only the medoids shown), "hist" (histogram of the distribution of the distances) or "gmm" (histogram of the distances overlayed with the Gaussian distributions)}

\item{layout}{The layout style of the tree, either "circular" (default), "radial", "slanted", "linear" or "rectangular" ("linear" or "rectangular" are the same).}

\item{...}{ignored}
}
\value{
A ggplot object containing the plot. It can be printed to standard output or saved using ggsave.
}
\description{
plots a class GGRaSP variable either the full tree, a reduced tree, or the gmm.
}
