% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.R
\name{nest1_.Spatial}
\alias{nest1_.Spatial}
\alias{nest_.Spatial}
\title{Nested Spatial}
\usage{
nest1_.Spatial(data, ...)

nest_.Spatial(data, ...)
}
\arguments{
\item{data}{sp Spatial object}

\item{...}{ignored}
}
\value{
nested tibble
}
\description{
Create nested tables from Spatial classes.
}
\details{
For \code{nest_} the tables are doubly nested.
For \code{nest1_} the tables are singly nested.
}

