% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.ggrasp}
\alias{summary.ggrasp}
\title{summary.ggrasp}
\usage{
\method{summary}{ggrasp}(object, ...)
}
\arguments{
\item{object}{ggrasp-class object}

\item{...}{ignored}
}
\description{
prints a summary of the class GGRaSP variable. Output includes medoids and cutoff value after the clustering
}
